+++ packages/web3-utils/src/errors.ts
@@ -149,9 +149,3 @@ export class InvalidCharCodeError extends Web3Error {
 		super(value, 'invalid char code given');
 	}
 }
-
-export class InvalidTypeAbiInput extends Web3Error {
-	public constructor(value: string) {
-		super(value, 'components found but type is not tuple');
-	}
-}
