+++ packages/web3-utils/src/index.ts
@@ -3,5 +3,4 @@ export * from './errors';
 export * from './validation';
 export * from './types';
 export * from './hash';
-export * from './abi';
 export * from './string_manipulation';
