+++ packages/web3-utils/test/fixtures/converters.ts
@@ -251,168 +251,3 @@ export const toWeiInvalidData: [[any, any], string][] = [
 	[['data', 'kwei'], 'Invalid value given "data". Error: not a valid number.'],
 	[['1234', 'uwei'], 'Invalid value given "uwei". Error: invalid unit.'],
 ];
-
-export const jsonInterfaceValidData: [any, string][] = [
-	[
-		{
-			// testing function type
-			name: 'myMethod',
-			type: 'function',
-			inputs: [
-				{
-					type: 'uint256',
-					name: 'myNumber',
-				},
-				{
-					type: 'string',
-					name: 'myString',
-				},
-			],
-		},
-		'myMethod(uint256,string)',
-	],
-	[
-		{
-			name: 'f',
-			type: 'function',
-			inputs: [
-				{
-					name: 's',
-					type: 'tuple',
-					components: [
-						{
-							name: 'a',
-							type: 'uint256',
-						},
-						{
-							name: 'b',
-							type: 'uint256[]',
-						},
-						{
-							name: 'c',
-							type: 'tuple[]',
-							components: [
-								{
-									name: 'x',
-									type: 'uint256',
-								},
-								{
-									name: 'y',
-									type: 'uint256',
-								},
-							],
-						},
-					],
-				},
-				{
-					name: 't',
-					type: 'tuple',
-					components: [
-						{
-							name: 'x',
-							type: 'uint256',
-						},
-						{
-							name: 'y',
-							type: 'uint256',
-						},
-					],
-				},
-				{
-					name: 'a',
-					type: 'uint256',
-				},
-			],
-			outputs: [],
-		},
-		'f((uint256,uint256[],(uint256,uint256)[]),(uint256,uint256),uint256)',
-	],
-	[
-		// testing event type
-		{
-			type: 'event',
-			inputs: [
-				{ name: 'a', type: 'uint256', indexed: true },
-				{ name: 'b', type: 'bytes32', indexed: false },
-			],
-			name: 'Event',
-		},
-		'Event(uint256,bytes32)',
-	],
-	[
-		{
-			name: 'myEvent',
-			type: 'event',
-			inputs: [
-				{
-					type: 'uint256',
-					name: 'myNumber',
-				},
-				{
-					type: 'bytes32',
-					name: 'myBytes',
-				},
-			],
-		},
-		'myEvent(uint256,bytes32)',
-	],
-];
-
-export const jsonInterfaceInvalidData: [any, string][] = [
-	[
-		{
-			name: 'f',
-			type: 'function',
-			inputs: [
-				{
-					name: 's',
-					type: 'notTuple',
-					components: [
-						{
-							name: 'a',
-							type: 'uint256',
-						},
-						{
-							name: 'b',
-							type: 'uint256[]',
-						},
-						{
-							name: 'c',
-							type: 'tuple[]',
-							components: [
-								{
-									name: 'x',
-									type: 'uint256',
-								},
-								{
-									name: 'y',
-									type: 'uint256',
-								},
-							],
-						},
-					],
-				},
-				{
-					name: 't',
-					type: 'tuple',
-					components: [
-						{
-							name: 'x',
-							type: 'uint256',
-						},
-						{
-							name: 'y',
-							type: 'uint256',
-						},
-					],
-				},
-				{
-					name: 'a',
-					type: 'uint256',
-				},
-			],
-			outputs: [],
-		},
-		'Invalid value given "notTuple". Error: components found but type is not tuple.',
-	],
-];
