+++ packages/web3-utils/test/unit/converters.test.ts
@@ -20,7 +20,6 @@ import {
 	toUtf8,
 	toWei,
 	utf8ToHex,
-	jsonInterfaceMethodToString,
 } from '../../src/converters';
 import {
 	asciiToHexValidData,
@@ -42,8 +41,6 @@ import {
 	toWeiValidData,
 	utf8ToHexInvalidData,
 	utf8ToHexValidData,
-	jsonInterfaceValidData,
-	jsonInterfaceInvalidData,
 } from '../fixtures/converters';
 
 describe('converters', () => {
@@ -321,30 +318,6 @@ describe('converters', () => {
 		describe('invalid cases', () => {
 			it.each(toWeiInvalidData)('%s', (input, output) => {
 				expect(() => toWei(input[0], input[1])).toThrow(output);
-			});
-		});
-	});
-
-	describe('toChecksumAddress', () => {
-		describe('valid cases', () => {
-			it.each(jsonInterfaceValidData)('%s', (input, output) => {
-				expect(jsonInterfaceMethodToString(input)).toEqual(output);
-			});
-		});
-		describe('invalid cases', () => {
-			// TODO: To be done after `sha3` is implemented
-			it.todo('should throw error for invalid cases');
-		});
-	});
-	describe('jsonInterface', () => {
-		describe('valid cases', () => {
-			it.each(jsonInterfaceValidData)('%s', (input, output) => {
-				expect(jsonInterfaceMethodToString(input)).toEqual(output);
-			});
-		});
-		describe('invalid cases', () => {
-			it.each(jsonInterfaceInvalidData)('%s', (input, output) => {
-				expect(() => jsonInterfaceMethodToString(input)).toThrow(output);
 			});
 		});
 	});
